# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hotel management API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/room": {
            "get": {
                "tags": [
                    "Room"
                ],
                "summary": "Get all rooms",
                "description": "List of rooms",
                "operationId": "1758cfa0a4f31164f32a8c2192f97d86",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Room"
                ],
                "summary": "Create a new room",
                "description": "Create a new room with the provided details",
                "operationId": "9be38be1eed9b9d7e62fd45fca294b7c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "number",
                                    "type_id",
                                    "hotel_id"
                                ],
                                "properties": {
                                    "number": {
                                        "type": "integer",
                                        "example": "101"
                                    },
                                    "type_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "hotel_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Room created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/room/{id}": {
            "get": {
                "tags": [
                    "Room"
                ],
                "summary": "Get a specific room",
                "description": "Retrieve a specific room by ID",
                "operationId": "c45e99f9ccf62d45114462c46ea5036a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the room",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Room not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Room"
                ],
                "summary": "Update an existing room",
                "description": "Update an existing room by ID",
                "operationId": "096109bd729e9ee91d620d1b883d87df",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the room",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "number",
                                    "type_id",
                                    "hotel_id"
                                ],
                                "properties": {
                                    "number": {
                                        "type": "integer",
                                        "example": "101"
                                    },
                                    "type_id": {
                                        "type": "integer",
                                        "example": "3"
                                    },
                                    "hotel_id": {
                                        "type": "integer",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Room updated successfully"
                    },
                    "404": {
                        "description": "Room not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Room"
                ],
                "summary": "Delete a specific room",
                "description": "Delete a specific room by ID",
                "operationId": "ba010fe020114b40d872de5e7913a8d5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the room",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Room deleted successfully"
                    },
                    "404": {
                        "description": "Room not found"
                    }
                }
            }
        },
        "/api/room/search/{number}": {
            "get": {
                "tags": [
                    "Room"
                ],
                "summary": "Search rooms by number",
                "description": "Search rooms by number",
                "operationId": "58770128b99b5deb310cc0834add642b",
                "parameters": [
                    {
                        "name": "number",
                        "in": "path",
                        "description": "Number of the room",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/room/{id}/type": {
            "get": {
                "tags": [
                    "Room"
                ],
                "summary": "Get types of a specific room",
                "description": "Get types of a specific room by ID",
                "operationId": "d32c0a33bb5094e496009174fe8ee3dc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the room",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Room not found"
                    }
                }
            }
        },
        "/api/room/{id}/hotels": {
            "get": {
                "tags": [
                    "Room"
                ],
                "summary": "Get hotels of a specific room",
                "description": "Get hotels of a specific room by ID",
                "operationId": "ec2ab8473e3ef4a3d912e9696abda626",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the room",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Room not found"
                    }
                }
            }
        },
        "/api/room/{id}/guests": {
            "get": {
                "tags": [
                    "Room"
                ],
                "summary": "Get guests of a specific room",
                "description": "Get guests of a specific room by ID",
                "operationId": "9852f7985d39209ab6aea43425d08c71",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the room",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "room not found"
                    }
                }
            }
        },
        "/api/room/{room}/guests": {
            "post": {
                "tags": [
                    "Room"
                ],
                "summary": "Add guests to a specific room",
                "description": "Add guests to a specific room by ID",
                "operationId": "ef88ca290aba5589e50429f49b7bb127",
                "parameters": [
                    {
                        "name": "room",
                        "in": "path",
                        "description": "ID of the room",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "room_id",
                                    "guest_id"
                                ],
                                "properties": {
                                    "room_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "guest_id": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Guests added successfully"
                    },
                    "404": {
                        "description": "room not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Room"
                ],
                "summary": "Remove guests from a specific room",
                "description": "Remove guests from a specific room by ID",
                "operationId": "9b7a39939fdad92915568e3398db2a3b",
                "parameters": [
                    {
                        "name": "room",
                        "in": "path",
                        "description": "ID of the room",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "room_id",
                                    "guest_id"
                                ],
                                "properties": {
                                    "room_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "guest_id": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Guests removed successfully"
                    },
                    "404": {
                        "description": "room not found"
                    }
                }
            }
        },
...
    },
...
}